refactor: Popup page Compound 패턴 제거
