+++ package.json
@@ -81,7 +81,6 @@
     "style-loader": "^3.3.1",
     "svgo-loader": "^4.0.0",
     "terser-webpack-plugin": "^5.3.6",
-    "ts-loader": "^9.4.1",
     "tsconfig-paths-webpack-plugin": "^4.0.0",
     "webpack": "^5.74.0",
     "webpack-bundle-analyzer": "^4.8.0",
@@ -92,7 +91,6 @@
   },
   "dependencies": {
     "@storybook/addon-actions": "^6.5.16",
-    "@tanstack/react-query": "^4.24.6",
     "@types/estree": "^1.0.0",
     "bestzip": "^2.2.1",
     "chart.js": "^3.9.1",
@@ -102,8 +100,6 @@
     "react": "^18.2.0",
     "react-chartjs-2": "^4.3.1",
     "react-dom": "^18.2.0",
-    "react-query": "^3.39.3",
-    "react-router-dom": "^6.8.1",
     "react-syntax-highlighter": "^15.5.0",
     "react-uid": "^2.3.2",
     "react-use": "^17.4.0",
