+++ src/components/shared/select/PartTitleSelect.tsx
@@ -15,9 +15,11 @@ const PartTitleSelect = ({
   const [isOpen, setIsOpen] = React.useState(false);
   const [selected, setSelected] = useRecoilState(problemTitleOption);
 
-  if (selected === 'ALL') {
-    setSelected(`전체 문제 (${allSolvedCnt})`);
-  }
+  React.useEffect(() => {
+    if (selected === 'ALL') {
+      setSelected(`전체 문제 (${allSolvedCnt})`);
+    }
+  }, []);
 
   const onChangePartTitle = React.useCallback((option: string) => {
     onChangePageIdx(0);
