+++ src/service/solution/index.ts
@@ -1,0 +1,4 @@
+import { filteredSolutions } from './filteredSolutions';
+import { getAllSolutions } from './getAllSolutions';
+
+export { filteredSolutions, getAllSolutions };
